+++ frontend/wailsjs/go/main/App.d.ts
@@ -8,7 +8,7 @@ export function ClearRecentFiles():Promise<void>;
 
 export function ExportContentToPDF(arg1:string):Promise<void>;
 
-export function ExportToHTML(arg1:string,arg2:string):Promise<void>;
+export function ExportToHTML(arg1:string):Promise<void>;
 
 export function ExportToPDF(arg1:string):Promise<void>;